Cleanup and restructuring in multiple files
